start moving BTI (after BTI refactor)
